refactor : IDE-117 파일 삭제 함수 커스텀 hook 분리
